Additional formatting for faceting. Plots closer together. Rehide "hanging" wrapped facet plot. Moved the plugins to their own file. Added more doc strings. Added title to grid in case it is made to render it.
